requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
marshmallow==3.20.1
bcrypt==4.0.1
alembic==1.12.0
prometheus-flask-exporter==0.23.0
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
flake8==6.1.0
black==23.9.1